Collect extracted mgWebGL collaborator modules under mgWebGLParts/
Pure relocation to make the refactor legible for review: the six modules
chipped out of MGWebGL (camera, viewTransforms, framebuffers,
canvasTextures, postProcessUniformBuffers, eventHandlers) move into a
mgWebGLParts/ subfolder, signalling they are private collaborators of the
one class. They have no importers outside mgWebGL.tsx, so blast radius is
just the relative-import depths (./ -> ../, ../ -> ../../) and the six
import lines in mgWebGL.tsx. No code content changed; git tracks all six
as renames. Project typechecks clean.
